pyproject.toml
[tool.poetry]
name = "clpsit"
version = "0.1.0"
description = ""
authors = ["teismar <eismarti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
ruff = "^0.2.1"
poethepoet = "^0.24.4"
bandit = "^1.7.7"
mypy = "^1.8.0"
safety = "^3.0.1"
dodgy = "^0.2.1"
semgrep = "^1.61.1"
detect-secrets = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W", "R", "C", "N", "S", "T"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
strict = true

[tool.bandit]
targets = ["."]
level = 3
confidence = 3

[tool.poe.tasks]
lint_black = "black . --check"
lint_isort = "isort . --check"
lint_ruff = "ruff ."
lint_mypy = "mypy ."
lint = ["lint_black", "lint_isort", "lint_ruff", "lint_mypy"]

security_bandit = "bandit -r ."
security_safety = "safety check"
security_dodgy = "dodgy"
security_semgrep = "semgrep --config=p/python --config=p/flask ."
security_detect_secrets = "detect-secrets scan && detect-secrets audit"
security = ["security_bandit", "security_safety", "security_dodgy", "security_semgrep"]

format_black = "black ."
format_isort = "isort ."
format_ruff = "ruff . --fix"
format = ["format_black", "format_isort", "format_ruff"]

all = ["lint", "security", "format"]